Renamed import of the TimeTimer in RepairToolTask and in GetToXZWithElytraTask
